Combine test cases
