+++ tests/functional/ext/no_self_use/no_self_use.txt
@@ -1,2 +1,2 @@
-no-self-use:16:4:16:23:Toto.function_method:Method could be a function:UNDEFINED
-no-self-use:24:4:24:35:Toto.async_function_method:Method could be a function:UNDEFINED
+no-self-use:14:4:14:23:Toto.function_method:Method could be a function:UNDEFINED
+no-self-use:22:4:22:35:Toto.async_function_method:Method could be a function:UNDEFINED
