--- tests/functional/ext/no_self_use/no_self_use_py3.py
@@ -1,12 +1,0 @@
-# pylint: disable=missing-docstring,unused-argument,invalid-name,too-few-public-methods
-
-class A:
-    def __init__(self):
-        self.store = {}
-
-    def get(self, key, default=None):
-        return self.store.get(key, default)
-
-class B(A):
-    def get_memo(self, obj):
-        return super().get(obj)
