--- tests/functional/ext/no_self_use/no_self_use_py3.rc
@@ -1,2 +1,0 @@
-[master]
-load-plugins=pylint.extensions.no_self_use
